Ok - since I'm in the mood for porting things, let's get the MACOS port going at the same time. This has the added advantage of finally cleaning up more of the argp stuff. What a mess.

+++ Makefile
@@ -11,8 +11,13 @@ ifdef _SOLARIS
     RM = rm -f
     CFLAGS += -DSOLARIS
     CXXFLAGS += -DSOLARIS
-else
-    RM = rm -fv 
+else 
+    ifdef _MACOS
+    	CFLAGS+= -DMACOS
+    	CXXFLAGS+= -DMACOS
+    else
+    	RM = rm -fv
+    endif 
 endif
 
 %: %.subst
